% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsCausal.R
\name{ctsCausalTest}
\alias{ctsCausalTest}
\title{Nonparametric test of flatness of a causal dose-response curve under exogeneity}
\usage{
ctsCausalTest(
  Y,
  A,
  W,
  method,
  conf.level = 0.95,
  dist = "TwoPoint",
  cross.fit = TRUE,
  SL.library = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of observed outcome values.}

\item{A}{\code{n x 1} numeric vector of exposure values.}

\item{W}{\code{n x p} data.frame of potential confounders.}

\item{method}{decides which method to be used for the test. Current supported methods are: \code{'isoreg'}, the method by (Westling et al., 2020); \code{'continuous'}, the method by (Weng et al., 2022) for a continuous treatment curve.}

\item{conf.level}{the confidence level of the test Default set to 0.95.}

\item{dist}{the distribution used by (Weng et al., 2022) for the test. It has to be either \code{"TwoPoint"} or \code{"Rademachar"}.}

\item{cross.fit}{boolean variable that controls whether cross fitting will be used for the estimation of nuisance parameters}

\item{verbose}{if set to TRUE, the function will output the progression of estimation and inferece}

\item{SL.libraries}{the libraries used for Superlearner. If set to NULL, then the default libraries \code{c("SL.mean", "SL.glm", "SL.gam", "SL.earth")} will be used.}
}
\value{
\code{causalNullTest} returns a list with the following elements:
\item{p.value}{The p value of the test.}
\item{test.stat}{Test statistic.}
If \code{method == "isoreg"}, the following elements are also included in the output:
\item{stat.ci.ll}{The lower limit of the confidence interval for the test statistic.}
\item{stat.ci.ul}{The upper limit of the confidence interval for the test statistic.}
}
\description{
This function performs a hypothesis test that the causal dose-response curve theta(a) is flat on the support of the observed exposure A. It includes two different method for the test. The method by (Westling et al., 2020) allows the exposure may be discrete, continuous, or an arbitrary mixture of discrete and continuous components, while the method by (Weng et al., 2022) only considers the continuous treatment.  See the accompanying paper for details.
}
\examples{
# Sample data
n <- 1000
W <- data.frame(W1 = runif(n))
Z <- rbinom(n, size = 1, prob = 1/(1 + exp(2-W$W1)))
A <- (1-Z) * rnorm(n, mean = W$W1, sd = abs(1 + W$W1))
Y <- rexp(n, rate = 1+abs(W$W1 * A))
ctsCausalTest(Y, A, W, method='isoreg', cross.fit=T, verbose=TRUE)
}
